+++ CHANGES.md
@@ -3,6 +3,8 @@ Change Log
 
 ### changes in 1.0-SNAPSHOT
 
+* changed X509V3ExtenstionsType inhibit_any_policy to Option[String]
+* changed NetworkTraffic src_port and dst_port to Option[Int]
 
 
 ### changes in 0.9